+++ packages/react-router-dev/CHANGELOG.md
@@ -1,5 +1,14 @@
 # `@react-router/dev`
 
+## 7.5.3
+
+### Patch Changes
+
+- Updated dependencies:
+  - `react-router@7.5.3`
+  - `@react-router/node@7.5.3`
+  - `@react-router/serve@7.5.3`
+
 ## 7.5.2
 
 ### Patch Changes